ros-setup/manager/requirements.txt
networkx
matplotlib
numpy
spacy
pandas
pyttsx3
SpeechRecognition
sounddevice
plac
pyyaml
rospkg
scipy